+++ scripts/uni_00_x=delta_y=eta_v=n_b_diff_stationary.py
@@ -1,23 +1,22 @@
 # dependencies
-import numpy as np
 import os 
 import sys
 
 # qom modules
-from qom.ui import init_log
-from qom.loopers import XYLooper
+from qom.utils.wrappers import wrap_looper
 
 # add path to local libraries
 sys.path.append(os.path.abspath(os.path.join('..', 'sync_bi_uni')))
 # import system
-from systems.Uni02 import Uni02
+from systems.Uni00 import Uni00
 
 # all parameters
 params = {
     'looper': {
+        'mode': 'multithread',
         'X': {
             'var': 'delta',
-            'min': 0.0,
+            'min': 0.00,
             'max': 0.02,
             'dim': 101
         },
@@ -27,20 +26,6 @@ params = {
             'max': 1.0,
             'dim': 101
         }
-    },
-    'solver': {
-        'show_progress': True,
-        'cache': False,
-        'cache_dir': 'H:/Workspace/VSCode/Python/sync_bi_uni/data/uni_02/0.0_10000.0_100001',
-        'method': 'RK45',
-        'measure_type': 'corr_ele',
-        'idx_row': [0, 1, 2, 3],
-        'idx_col': [0, 1, 2, 3],
-        'range_min': 0,
-        'range_max': 1001,
-        't_min': 0,
-        't_max': 100,
-        't_dim': 1001
     },
     'system': {
         'A_l': 52.0,
@@ -55,45 +40,32 @@ params = {
     },
     'plotter': {
         'type': 'contourf',
-        'palette': 'Reds',
+        'palette': 'blr',
         'bins': 11,
         'x_label': '$\\delta / \\omega_{mL}$',
         'x_bound': 'both',
-        'x_ticks': [0.0, 0.01, 0.02],
+        'x_ticks': [0.00, 0.01, 0.02],
         'y_label': '$\\eta$',
         'y_bound': 'both',
         'y_ticks': [0.5, 0.75, 1.0],
-        'cbar_title': '$n_{n_{2}} - n_{n_{1}}$',
-        'cbar_ticks': [- 0.007119, - 0.007117, - 0.007115],
+        'cbar_title': '$n_{b_{R}} - n_{b_{L}}$',
+        'cbar_ticks': [0.006, 0.012, 0.018],
         'cbar_position': 'top',
         'label_font_size': 22,
         'tick_font_size': 18
     }
 }
 
-# initialize system
-system = Uni02(params['system'])
-
-# get steady state values
-modes, corrs = system.get_modes_corrs_stationary(system.get_mode_rates, system.get_ivc, system.get_A)
-# difference
-print((corrs[3][3] + corrs[2][2] - corrs[1][1] - corrs[0][0]) / 2)
-
-# initialize log
-init_log()
-
-# initialize system
-system = Uni02(params['system'])
-
 # function to obtain the difference between the phonon numbers
 def func_n_b_diff(system_params, val, logger, results):
     # update system
-    system.params = system_params
-    # get difference
-    n_b_diff = system.get_n_b_diff()
+    system = Uni00(system_params)
+    # get steady state values
+    _, corrs = system.get_modes_corrs_stationary(system.get_mode_rates, system.get_ivc, system.get_A)
+    # calculate difference
+    n_b_diff = (corrs[7][7] + corrs[6][6] - corrs[3][3] - corrs[2][2]) / 2
     # update results
     results.append((val, n_b_diff))
 
 # looper 
-looper = XYLooper(func_n_b_diff, params)
-looper.loop(plot=True)
+looper = wrap_looper(Uni00, params, func_n_b_diff, 'XYLooper', plot=True)